Move `NodeData` struct into `taffy_tree` module
This struct is only used in the taffy_tree module, and can be made
private if it also defined within this module.
