+++ src/tree/node.rs
@@ -1,9 +1,6 @@
 //! UI node types and related data structures.
 //!
 //! Layouts are composed of multiple nodes, which live in a tree-like data structure.
-use crate::style::Style;
-use crate::tree::Cache;
-use crate::tree::Layout;
 
 #[cfg(feature = "taffy_tree")]
 use slotmap::{DefaultKey, Key, KeyData};
@@ -59,49 +56,5 @@ impl From<NodeId> for DefaultKey {
     #[inline]
     fn from(key: NodeId) -> Self {
         KeyData::from_ffi(key.0).into()
-    }
-}
-
-/// Layout information for a given [`Node`](crate::node::Node)
-///
-/// Stored in a [`TaffyTree`].
-pub(crate) struct NodeData {
-    /// The layout strategy used by this node
-    pub(crate) style: Style,
-
-    /// The always unrounded results of the layout computation. We must store this separately from the rounded
-    /// layout to avoid errors from rounding already-rounded values. See <https://github.com/DioxusLabs/taffy/issues/501>.
-    pub(crate) unrounded_layout: Layout,
-
-    /// The final results of the layout computation.
-    /// These may be rounded or unrounded depending on what the `use_rounding` config setting is set to.
-    pub(crate) final_layout: Layout,
-
-    /// Should we try and measure this node?
-    pub(crate) needs_measure: bool,
-
-    /// The cached results of the layout computation
-    pub(crate) cache: Cache,
-}
-
-impl NodeData {
-    /// Create the data for a new node
-    #[must_use]
-    pub const fn new(style: Style) -> Self {
-        Self {
-            style,
-            cache: Cache::new(),
-            unrounded_layout: Layout::new(),
-            final_layout: Layout::new(),
-            needs_measure: false,
-        }
-    }
-
-    /// Marks a node and all of its parents (recursively) as dirty
-    ///
-    /// This clears any cached data and signals that the data must be recomputed.
-    #[inline]
-    pub fn mark_dirty(&mut self) {
-        self.cache.clear()
     }
 }
